# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SignalGen Go API
  version: 0.9.0
  description: |
    Implemented API surface for the incremental SignalGen Go backend.
    The Go API performs email/password and recovery flows through Supabase Auth,
    verifies returned bearer tokens, and issues a separate opaque app-session
    token. Proposed routes that are not implemented are intentionally excluded.
servers:
  - url: http://127.0.0.1:8080
    description: Local Docker development
...
paths:
...
  /api/v1/stocks:
    get:
      tags: [Analysis]
      operationId: listStockCatalog
      summary: List the three server-approved IDX instruments
      security:
        - supabaseBearer: []
          appSession: []
      responses:
        "200":
          description: Stock catalog.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/StockInstrument"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
  /api/v1/stock-universes:
    get:
      tags: [Analysis]
      operationId: listStockUniverses
      summary: List stock universes owned by the current user
      security:
        - supabaseBearer: []
          appSession: []
      responses:
        "200":
          description: Owner-scoped universe list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/StockUniverse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    post:
      tags: [Analysis]
      operationId: createStockUniverse
      summary: Create an owner-scoped stock universe
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StockUniverseWrite"}
      responses:
        "201":
          description: Universe created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StockUniverse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/InvalidRequest"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
  /api/v1/stock-universes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get:
      tags: [Analysis]
      operationId: getStockUniverse
      summary: Read an owner-scoped stock universe
      security:
        - supabaseBearer: []
          appSession: []
      responses:
        "200":
          description: Universe detail.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StockUniverse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Analysis]
      operationId: updateStockUniverse
      summary: Replace name and members using optimistic versioning
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: "#/components/schemas/StockUniverseWrite"}
                - type: object
                  required: [version]
                  properties:
                    version: {type: integer, minimum: 1}
      responses:
        "200":
          description: Universe updated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StockUniverse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {description: Version conflict.}
        "422": {$ref: "#/components/responses/InvalidRequest"}
    delete:
      tags: [Analysis]
      operationId: deleteStockUniverse
      summary: Delete an owner-scoped stock universe
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: {type: integer, minimum: 1}
      responses:
        "204": {description: Universe deleted.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {description: Version conflict.}
...
  /api/v1/subscription/plans:
    get:
      tags: [Subscription]
      operationId: listSubscriptionPlans
      summary: List active product plans and their feature mapping
      description: Prices and checkout are intentionally excluded until commercial terms and a payment provider are approved.
      security: []
      responses:
        "200":
          description: Active plans. This response does not activate access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlanList"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/subscription/cancel:
    post:
      tags: [Subscription]
      operationId: cancelCurrentSubscription
      summary: Cancel the current account subscription now or at period end
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCancelRequest"
      responses:
        "200":
          description: Updated owner-scoped subscription.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subscription is already canceled or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/operator/subscriptions:
    post:
      tags: [Operator]
      operationId: activateOperatorSubscription
      summary: Activate or change a subscription through a trusted operator
      description: Temporary MVP path. The actor is derived from authenticated operator context; browser checkout data cannot activate access.
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorSubscriptionActivateRequest"
      responses:
        "201":
          description: Subscription activated or changed with an append-only event.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued and refreshed by Supabase Auth.
    appSession:
      type: apiKey
      in: header
      name: X-App-Session
      description: Opaque one-time-issued SignalGen app-session secret.
  headers:
    PrivateNoStore:
      description: Protected responses must not be stored by shared or browser caches.
      schema:
        type: string
        example: private, no-store
...
  responses:
    InvalidRequest:
      description: Request syntax, values, or supported capability is invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Bearer or app-session credential is missing or invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Session, account, or requested access is inactive.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource is missing or not visible to the current owner.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PayloadTooLarge:
      description: JSON request body exceeds the 64 KiB API limit.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: {code: PAYLOAD_TOO_LARGE, message: Body JSON melebihi batas 64 KiB., request_id: req_opaque}
    RateLimited:
      description: The authenticated account exceeded the mutation limit for this operation group.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
        Retry-After:
          description: Whole seconds until the current rate-limit window resets.
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: {code: RATE_LIMITED, message: Terlalu banyak permintaan. Coba lagi nanti., request_id: req_opaque}
    ServiceUnavailable:
      description: A required local or upstream dependency is unavailable.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string, pattern: "^req_"}
...
    SubscriptionPlan:
      type: object
      required: [code, name, description, features]
      properties:
        code: {type: string, enum: [free, analyst, pro]}
        name: {type: string}
        description: {type: string}
        features:
          type: array
          items: {type: string, enum: [screener, backtest]}
    SubscriptionPlanList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionPlan"
    Subscription:
      type: object
      required: [id, plan_code, plan_name, status, features, current_period_start, current_period_end, cancel_at_period_end, source, created_at, updated_at]
      properties:
        id: {type: string, pattern: "^sub_"}
        user_id:
          type: string
          description: Returned only by the operator mutation response.
        plan_code: {type: string, enum: [free, analyst, pro]}
        plan_name: {type: string}
        status: {type: string, enum: [trialing, active, past_due, canceled, expired]}
        features:
          type: array
          items: {type: string, enum: [screener, backtest]}
        current_period_start: {type: string, format: date-time}
        current_period_end: {type: string, format: date-time}
        cancel_at_period_end: {type: boolean}
        source: {type: string, enum: [manual, payment_provider]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    SubscriptionResponse:
      type: object
      required: [subscription]
      properties:
        subscription:
          $ref: "#/components/schemas/Subscription"
    SubscriptionCancelRequest:
      type: object
      additionalProperties: false
      required: [at_period_end]
      properties:
        at_period_end: {type: boolean}
        reason: {type: string, maxLength: 500}
    OperatorSubscriptionActivateRequest:
      type: object
      additionalProperties: false
      required: [user_id, plan_code, current_period_end, reason]
      properties:
        user_id: {type: string, minLength: 1}
        plan_code: {type: string, enum: [free, analyst, pro]}
        current_period_end: {type: string, format: date-time}
        reason: {type: string, minLength: 1, maxLength: 500}
...
    StockInstrument:
      type: object
      required: [symbol, name, exchange, currency]
      properties:
        symbol: {type: string, enum: [BBCA.JK, BBRI.JK, TLKM.JK]}
        name: {type: string}
        exchange: {type: string, enum: [XIDX]}
        currency: {type: string, enum: [IDR]}
    StockUniverseWrite:
      type: object
      additionalProperties: false
      required: [name, symbols]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        symbols:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items: {type: string, enum: [BBCA.JK, BBRI.JK, TLKM.JK]}
    StockUniverse:
      allOf:
        - $ref: "#/components/schemas/StockUniverseWrite"
        - type: object
          required: [id, version, created_at, updated_at]
          properties:
            id: {type: string, pattern: "^univ_"}
            version: {type: integer, minimum: 1}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
...